rename files to contain dashes
