+++ package.json
@@ -14,7 +14,7 @@
 		"start": "node index.js",
 		"test": "NODE_ENV=test mocha --require esm --exit --colors --bail tests/*",
 		"test:fast": "NODE_ENV=test mocha --exit --colors --bail tests/*puppeteer*.*",
-		"test:ci": "NODE_ENV=test mocha  --check-leaks --exit \"tests/!(*botbase*|*helper_puppeteer*).js\" --reporter mocha-junit-reporter --reporter-options mochaFile=./test-results/mocha/results.xml",
+		"test:ci": "NODE_ENV=test mocha  --check-leaks --exit \"tests/!(*botbase*|*helper-puppeteer*).js\" --reporter mocha-junit-reporter --reporter-options mochaFile=./test-results/mocha/results.xml",
 		"lint": "eslint . --ext .js,.ts",
 		"lint:fix": "npm run lint -- --fix"
 	},
